Refactor source to be in different modules
