+++ src/dom.ts
@@ -5,7 +5,7 @@ import {
 	ElementValue,
 	Portal,
 	Renderer,
-} from "./crank";
+} from "./core";
 
 const SVG_NAMESPACE = "http://www.w3.org/2000/svg";
