+++ umd.ts
@@ -1,4 +1,4 @@
 // input file for the UMD build
-export * from "./src/crank";
+export * from "./src/core/index";
 export * as dom from "./src/dom";
 export * as html from "./src/html";
